feat: Add watermark functionality
Add the ability to set a high watermark for the the ring buffer.

Whenever the ring buffer contains more than the watermark amount of
bytes, a pre-registered watermark callback is called on the publish()
 that pushes the buffer over the watermark.

That callback can then be used to efficiently signal out-of-band (for example,
via a eventfd) that some pre-determined amount of data is ready to read,
eliminating the need to busy wait for readability of the ring buffer
and allowing more efficient batched processing of ring buffer elements.